move sonar files into sonar dirs
