+++ integration_tests/sonar/test_sonar_django_json_response_type.py
