+++ integration_tests/sonar/test_sonar_exception_without_raise.py
