+++ integration_tests/sonar/test_sonar_fix_assert_tuple.py
