+++ integration_tests/sonar/test_sonar_fix_missing_self_or_cls.py
