+++ integration_tests/sonar/test_sonar_flask_json_response_type.py
