+++ integration_tests/sonar/test_sonar_jinja2_autoescape.py
