+++ integration_tests/sonar/test_sonar_remove_assertion_in_pytest_raises.py
