+++ src/core_codemods/sonar/sonar_django_json_response_type.py
@@ -1,5 +1,5 @@
-from codemodder.codemods.sonar import SonarCodemod
 from core_codemods.django_json_response_type import DjangoJsonResponseType
+from core_codemods.sonar.api import SonarCodemod
 
 SonarDjangoJsonResponseType = SonarCodemod.from_core_codemod(
     name="django-json-response-type-S5131",
