+++ src/core_codemods/sonar/sonar_enable_jinja2_autoescape.py
@@ -1,5 +1,5 @@
-from codemodder.codemods.sonar import SonarCodemod
 from core_codemods.enable_jinja2_autoescape import EnableJinja2Autoescape
+from core_codemods.sonar.api import SonarCodemod
 
 SonarEnableJinja2Autoescape = SonarCodemod.from_core_codemod(
     name="enable-jinja2-autoescape-S5247",
