+++ src/core_codemods/sonar/sonar_numpy_nan_equality.py
@@ -1,5 +1,5 @@
-from codemodder.codemods.sonar import SonarCodemod
 from core_codemods.numpy_nan_equality import NumpyNanEquality
+from core_codemods.sonar.api import SonarCodemod
 
 SonarNumpyNanEquality = SonarCodemod.from_core_codemod(
     name="numpy-nan-equality-S6725",
