+++ src/core_codemods/sonar/sonar_secure_random.py
@@ -1,5 +1,5 @@
-from codemodder.codemods.sonar import SonarCodemod
 from core_codemods.secure_random import SecureRandom
+from core_codemods.sonar.api import SonarCodemod
 
 SonarSecureRandom = SonarCodemod.from_core_codemod(
     name="secure-random-S2245",
