+++ tests/codemods/sonar/test_sonar_exception_without_raise.py
