+++ tests/codemods/sonar/test_sonar_fix_missing_self_or_cls.py
