+++ tests/codemods/sonar/test_sonar_literal_or_new_object_identity.py
