+++ tests/codemods/sonar/test_sonar_numpy_nan_equality.py
